requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
alembic==1.14.0
httpx==0.28.1
python-dateutil==2.9.0
redis==5.2.1
celery==5.4.0
openpyxl==3.1.5
reportlab==4.2.5
pytest==8.3.4
pytest-cov==6.0.0
